Automatically Committed for MakeGamesWithUs

+++ PeevedPenguins.spritebuilder/Gameplay.m
@@ -28,7 +28,7 @@ static const float MIN_SPEED = 5.f;
     int _currentTurn;
     
     // Time the bird has been idle
-    int _timeIdle;
+    float _timeIdle;
     
     // Following action, controls screen panning
     CCAction *_followPenguin;
@@ -60,7 +60,7 @@ static const float MIN_SPEED = 5.f;
     // Set current turn to 0
     _currentTurn = 0;
     
-    _timeIdle = 0;
+    _timeIdle = 0.f;
     
 }
 
@@ -71,12 +71,12 @@ static const float MIN_SPEED = 5.f;
     // if penguin slows to a critical point, allow next attempt
     if (ccpLength(_currentPenguin.physicsBody.velocity) < MIN_SPEED) { // ccpLength method takes the length of the velocity vector --> speed
         _timeIdle += (float)delta;
-        if (_timeIdle > 2) {
+        if (_timeIdle > 2.f) {
             [self nextAttempt];
             return;
         }
         
-    } else _timeIdle = 0;
+    } else _timeIdle = 0.f;
     
     int xMin = _currentPenguin.boundingBox.origin.x;
